requirements.txt
Django==1.11.6
django-appconf==1.0.2
django-compressor==2.2
django-constance==2.0.0
django-el-pagination==3.1.0
django-froala-editor==2.7.1
django-jet==1.0.6
django-modeltranslation==0.12.1
django-mptt==0.8.7
django-picklefield==1.0.0
django-taggit==0.22.1
gunicorn==19.4.5
olefile==0.44
Pillow==4.3.0
python-memcached==1.58
pytz==2017.2
rcssmin==1.0.6
rjsmin==1.0.12
six==1.11.0
sorl-thumbnail==12.3
whitenoise==3.3.1
